+++ src/main/java/gift/view/ProductViewController.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
