+++ src/test/java/gift/ProductServiceTest.java
@@ -5,6 +5,9 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import gift.domain.Product;
+import gift.repository.ProductDao;
+import gift.service.ProductService;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
